+++ packages/docusaurus-plugin-client-redirects/src/collectRedirects.ts
@@ -7,8 +7,11 @@
 
 import _ from 'lodash';
 import logger from '@docusaurus/logger';
-import {removeTrailingSlash} from '@docusaurus/utils';
-import {applyTrailingSlash, addTrailingSlash} from '@docusaurus/utils-common';
+import {
+  applyTrailingSlash,
+  addTrailingSlash,
+  removeTrailingSlash,
+} from '@docusaurus/utils-common';
 import {
   createFromExtensionsRedirects,
   createToExtensionsRedirects,
