+++ packages/docusaurus-plugin-content-docs/src/slug.ts
@@ -5,12 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {
-  addLeadingSlash,
-  isValidPathname,
-  resolvePathname,
-} from '@docusaurus/utils';
-import {addTrailingSlash} from '@docusaurus/utils-common';
+import {isValidPathname, resolvePathname} from '@docusaurus/utils';
+import {addLeadingSlash, addTrailingSlash} from '@docusaurus/utils-common';
 import {
   DefaultNumberPrefixParser,
   stripPathNumberPrefixes,
