+++ packages/docusaurus-utils-common/src/index.ts
@@ -12,6 +12,10 @@ export const blogPostContainerID = '__blog-post-container';
 export {
   default as applyTrailingSlash,
   addTrailingSlash,
+  addLeadingSlash,
+  addPrefix,
+  removeSuffix,
+  removeTrailingSlash,
   type ApplyTrailingSlashParams,
 } from './applyTrailingSlash';
 export {getErrorCausalChain} from './errorUtils';
