+++ packages/docusaurus-utils-validation/package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "@docusaurus/logger": "3.0.0",
     "@docusaurus/utils": "3.0.0",
+    "@docusaurus/utils-common": "3.0.0",
     "joi": "^17.9.2",
     "js-yaml": "^4.1.0",
     "tslib": "^2.6.0"
