+++ packages/docusaurus-utils/src/__tests__/jsUtils.test.ts
@@ -8,22 +8,10 @@
 import {jest} from '@jest/globals';
 import _ from 'lodash';
 import {
-  removeSuffix,
   removePrefix,
   mapAsyncSequential,
   findAsyncSequential,
 } from '../jsUtils';
-
-describe('removeSuffix', () => {
-  it("is no-op when suffix doesn't exist", () => {
-    expect(removeSuffix('abcdef', 'ijk')).toBe('abcdef');
-    expect(removeSuffix('abcdef', 'abc')).toBe('abcdef');
-    expect(removeSuffix('abcdef', '')).toBe('abcdef');
-  });
-  it('removes suffix', () => {
-    expect(removeSuffix('abcdef', 'ef')).toBe('abcd');
-  });
-});
 
 describe('removePrefix', () => {
   it("is no-op when prefix doesn't exist", () => {
