+++ packages/docusaurus-utils/src/__tests__/urlUtils.test.ts
@@ -5,14 +5,11 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {addTrailingSlash} from '@docusaurus/utils-common';
 import {
   normalizeUrl,
   getEditUrl,
   fileToPath,
   isValidPathname,
-  addLeadingSlash,
-  removeTrailingSlash,
   resolvePathname,
   encodePath,
   buildSshUrl,
@@ -204,33 +201,6 @@ describe('isValidPathname', () => {
     expect(isValidPathname('https://fb.com/hey')).toBe(false);
     expect(isValidPathname('//hey')).toBe(false);
     expect(isValidPathname('////')).toBe(false);
-  });
-});
-
-describe('addTrailingSlash', () => {
-  it('is no-op for path with trailing slash', () => {
-    expect(addTrailingSlash('/abcd/')).toBe('/abcd/');
-  });
-  it('adds / for path without trailing slash', () => {
-    expect(addTrailingSlash('/abcd')).toBe('/abcd/');
-  });
-});
-
-describe('addLeadingSlash', () => {
-  it('is no-op for path with leading slash', () => {
-    expect(addLeadingSlash('/abc')).toBe('/abc');
-  });
-  it('adds / for path without leading slash', () => {
-    expect(addLeadingSlash('abc')).toBe('/abc');
-  });
-});
-
-describe('removeTrailingSlash', () => {
-  it('is no-op for path without trailing slash', () => {
-    expect(removeTrailingSlash('/abcd')).toBe('/abcd');
-  });
-  it('removes / for path with trailing slash', () => {
-    expect(removeTrailingSlash('/abcd/')).toBe('/abcd');
   });
 });
