+++ packages/docusaurus/src/server/brokenLinks.ts
@@ -10,12 +10,11 @@ import logger from '@docusaurus/logger';
 import {matchRoutes as reactRouterMatchRoutes} from 'react-router-config';
 import {
   parseURLPath,
-  removeTrailingSlash,
   serializeURLPath,
   flattenRoutes,
   type URLPath,
 } from '@docusaurus/utils';
-import {addTrailingSlash} from '@docusaurus/utils-common';
+import {addTrailingSlash, removeTrailingSlash} from '@docusaurus/utils-common';
 import type {RouteConfig, ReportingSeverity} from '@docusaurus/types';
 
 function matchRoutes(routeConfig: RouteConfig[], pathname: string) {
