+++ packages/docusaurus/src/server/configValidation.ts
@@ -9,11 +9,13 @@ import {
   DEFAULT_PARSE_FRONT_MATTER,
   DEFAULT_STATIC_DIR_NAME,
   DEFAULT_I18N_DIR_NAME,
-  addLeadingSlash,
-  removeTrailingSlash,
 } from '@docusaurus/utils';
 import {Joi, printWarning} from '@docusaurus/utils-validation';
-import {addTrailingSlash} from '@docusaurus/utils-common';
+import {
+  addTrailingSlash,
+  addLeadingSlash,
+  removeTrailingSlash,
+} from '@docusaurus/utils-common';
 import type {
   DocusaurusConfig,
   I18nConfig,
